Prevent imports when running validation (#4)
* Prevent imports when running validation

pylint by default builds the AST of all the dependencies it can find
when running the py3k validation. In some cases (kubernetes), it can
take a while to run, while not providing value as we don't care about
checking attributes from those. This works around it by setting sys.path
just when running the linter, so that imports fail and validation runs
faster.

* Try a more surgical change